Refactor: moving model formatting to FormsService
