+++ frontend/src/app/modules/common/dynamic-forms/services/dynamic-fields/dynamic-fields.service.ts
@@ -9,6 +9,7 @@ import { Observable, of } from "rxjs";
 import { map } from "rxjs/operators";
 import { HttpClient } from "@angular/common/http";
 import { I18nService } from "core-app/modules/common/i18n/i18n.service";
+import { FormsService } from "core-app/core/services/forms/forms.service";
 
 
 @Injectable()
@@ -118,6 +119,7 @@ export class DynamicFieldsService {
   constructor(
     private httpClient:HttpClient,
     private I18n:I18nService,
+    private formsService:FormsService,
   ) {
   }
 
@@ -136,28 +138,7 @@ export class DynamicFieldsService {
   }
 
   getModel(formPayload:IOPFormModel):IOPFormModel {
-    return this.getFormattedFieldsModel(formPayload);
-  }
-
-  getFormattedFieldsModel(formModel:IOPFormModel = {}):IOPFormModel {
-    const { _links: resourcesModel, _meta: metaModel, ...otherElements } = formModel;
-    const otherElementsModel = Object.keys(otherElements).reduce((model, key) => {
-      const elementValue = otherElements[key];
-
-      if (this.isValue(elementValue)) {
-        model = {...model, [key]:elementValue}
-      }
-
-      return model;
-    }, {})
-
-    const model = {
-      ...otherElementsModel,
-      _meta: metaModel,
-      ...this.getFormattedResourcesModel(resourcesModel),
-    };
-
-    return model;
+    return this.formsService.formatModelToEdit(formPayload);
   }
 
   getFormlyFormWithFieldGroups(fieldGroups:IDynamicFieldGroupConfig[] = [], formFields:IOPFormlyFieldSettings[] = []):IOPFormlyFieldSettings[] {
@@ -191,24 +172,6 @@ export class DynamicFieldsService {
 
   private isFieldSchema(schemaValue:IOPFieldSchemaWithKey|any):boolean {
     return !!schemaValue?.type;
-  }
-
-  private getFormattedResourcesModel(resourcesModel:IOPFormModel['_links'] = {}):IOPFormModel['_links'] {
-    return Object.keys(resourcesModel).reduce((result, resourceKey) => {
-      const resource = resourcesModel[resourceKey];
-      // ng-select needs a 'name' in order to show the label
-      // We need to add it in case of the form payload (HalLinkSource)
-      const resourceModel = Array.isArray(resource) ?
-        resource.map(resourceElement => ({...resourceElement, name: resourceElement?.name || resourceElement?.title})) :
-        {...resource, name: resource?.name || resource?.title};
-
-      result = {
-        ...result,
-        ...this.isValue(resourceModel) && {[resourceKey]: resourceModel},
-      };
-
-      return result;
-    }, {});
   }
 
   private getFormlyFieldConfig(fieldSchema:IOPFieldSchemaWithKey, formPayload:IOPFormModel):IOPFormlyFieldSettings|null {
@@ -394,10 +357,6 @@ export class DynamicFieldsService {
 
   private isMultiSelectField(field:IOPFieldSchemaWithKey) {
     return field?.type?.startsWith('[]');
-  }
-
-  private isValue(value:any) {
-    return ![null, undefined, ''].includes(value);
   }
 }
